+++ src/WaterModels.jl
@@ -65,7 +65,7 @@ module WaterModels
 
     include("prob/wf.jl")
     include("prob/owf.jl")
-    include("prob/ne.jl")
+    include("prob/des.jl")
 
     include("core/export.jl")
 end