+++ test/runtests.jl
@@ -42,7 +42,7 @@ include("common.jl")
 
     include("owf.jl")
 
-    include("ne.jl")
+    include("des.jl")
 
     include("warm_start.jl")