+++ CHANGELOG.md
@@ -5,6 +5,7 @@ WaterModels.jl Change Log
 - Update to InfrastructureModels v0.4.3.
 - Remove the Constrained Water Flow (cwf) problem specification.
 - Rename `post_` methods to `solve_`.
+- Rename `ne` (network expansion) methods to `des` (design).
 
 ### v0.2.0
 - Update to JuMP v0.21 and InfrastructureModels v0.4.